WIP: Correct zone test to have trailing dot
PowerDNS expects that all zones have a trailing dot.
The test data did not have a zone in any of the test zones, which is
basically a grave mistake. The trailing dot handling is broken and
wasn't caught by the lack of the trailing dot in the zone data itself.